Add `page` parameter to /admin/users/list API
Related to https://github.com/discourse/discourse/pull/4866, document
the `page` parameter.

I also noted that several of the parameters were marked as in `path`,
but I believe these should be in `query`... if I'm wrong, and these
need to be `path` for some reason, just let me know and I'll roll them
back.  But, FWIW, when I try using the API myself, those values only
seem to work when in the querystring.