DEV: Remove lefhook validation
Since we aren't actually supposed to be editing the openapi.json file
directly in the repo we aren't using this lefthook validator.

It's always really strict at the moment and doesn't match the swagger
validator we just created a githun workflow for.

Let's remove this for now and later we can setup another github workflow
action to use the redocly validator that this lefthook was using.