DEV: Auto trigger validation workflow (#158)
To avoid infinite loops apparently github won't trigger fire a PR action
if a bot creates a PR. If I manually create a PR the validation action
will run, but not if the update_docs action creates it.

This change updates the validate_pr workflow to trigger on completion of
the update_docs workflow. Hopefully this works.